fix for issue #454
stroke-dasharray appeared to not work on path elements when the path drawn was a straight line.  This was caused by the call to ctx.closePath() which drew another line back to the origin on top of the original line.  The result looked like a solid line rather than a dashed line.

Fix is to only close the path if it is not a straight line.